Add an AuthenticatedQuery class
