+++ server/src/main/java/io/javaclasses/filehub/api/GetRootFolderId.java
@@ -3,6 +3,6 @@ package io.javaclasses.filehub.api;
 /**
  * A {@link Query} that represents intention of a client to get the identifier of the root folder.
  */
-public class GetRootFolderId extends Query {
+public class GetRootFolderId extends AuthenticatedQuery {
 
 }
