+++ server/src/main/java/io/javaclasses/filehub/api/GetUser.java
@@ -6,6 +6,6 @@ import com.google.errorprone.annotations.Immutable;
  * A {@link Query} that represents intention of a client to get the current user.
  */
 @Immutable
-public final class GetUser extends Query {
+public final class GetUser extends AuthenticatedQuery {
 
 }
